Split output-controlling flags out to their own struct to keep things cleaner when using Request.
Remove "-short" flag as there isn't a compelling use case for it I can
see right now, but it can be easily added back as an option of
printCommitMessages().

Explicitly declare the keys we're populating values for in
&github.ListOptions in case the provider's API changes in the future and
some values get flipped in a different order.

Return the struct slice []CommitMessage from GetCommitMessages() so it
can be used elsewhere for an easy time formatting pretty messages.

Add printCommitMessages() to handle pretty-printing stuff and output
control flags.

At this point, the tool basically functions, and will successfully fetch
commit messages and basic metadata between HEAD and BASE, and allow the
user to output in json or a more reader-friendly default format.