Add first pass tests.
Define a lot of boilerplate required to construct mock responses from
GitHub that GetCommitMessages can process.

Pagination is a real problem with this method, as the code was designed
to handle it from GitHub, but our mock CommitsComparison response just
sends the same amount of commits back each time, making tests break if
you define one over 100 commits in size (the maximum for a real GitHub
page, for which our code to be tested accounts). I'm thinking testing it
against the real API instead of mock is the only sane way to go forward.